integrate jpmml for area estimation.  Implement caching for exhaustive banking to hopefully speed up search.  Add flag to use either crude area estimation or the new ml-absed estimation.  integrate area in banking (with a few todos noted)

+++ src/spatial/metadata/memory/BankingData.scala
@@ -89,7 +89,7 @@ case class Instance(
   metapipe: Option[Ctrl],           // Controller if at least some accesses require n-buffering
   banking:  Seq[Banking],           // Banking information
   depth:    Int,                    // Depth of n-buffer
-  cost:     Long,                    // Cost estimate of this configuration
+  cost:     Double,                    // Cost estimate of this configuration
   ports:    Map[AccessMatrix,Port], // Buffer ports
   padding:  Seq[Int],               // Padding for memory based on banking
   darkVolume: Int,                  // Number of elements inaccessible due to B > 1
